dw2i: fix potential integer overflow.
Merge from ccid 1.4.34.

+++ src/ccid.h
@@ -360,7 +360,7 @@ void acr38_error(int error, const char *file, int line, const char *function);
 void EnablePicc(unsigned int reader_index, int enabled);
 
 /* convert a 4 byte integer in USB format into an int */
-#define dw2i(a, x) (unsigned int)((((((a[x+3] << 8) + a[x+2]) << 8) + a[x+1]) << 8) + a[x])
+#define dw2i(a, x) (unsigned int)(((((((unsigned int)a[x+3] << 8) + (unsigned int)a[x+2]) << 8) + (unsigned int)a[x+1]) << 8) + (unsigned int)a[x])
 
 /* all the data rates specified by ISO 7816-3 Fi/Di tables */
 #define ISO_DATA_RATES 10753, 14337, 15625, 17204, \
